fixed compil on 64 bit systems
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@45310 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsalgo/src/aimsalgo/geodesicpath/geodesic_algorithm_exact.h
@@ -143,7 +143,7 @@ private:
   double start[6];
   interval_pointer i_new[5];
 
-  unsigned m_queue_max_size;      //used for statistics
+  size_t m_queue_max_size;      //used for statistics
   unsigned m_iterations;      //used for statistics
 
   SortedSources m_sources;
